Module bones::ecs

Entity component system for the bones library.

Modules

atomic

Atomic Refcell implmentation.

bitset

Bitset implementation.

components

ECS component storage.

dispatcher

System executor for Worlds.

entities

Entity implementation, storage, and interation.

prelude

The prelude.

resources

World resource storage.

system

Implements the system API for the ECS.

uuid

UUID-related utilities such as UUID map and type UUIDs.

Macros

join

The join macro makes it very easy to iterate over multiple components of the same Entity at once.

Structs

World

The World is simply a collection of Resources, and ComponentStores.

Enums

EcsError

The types of errors used throughout the ECS.

Traits

EcsData

Helper trait that is auto-implemented for anything that may be stored in the ECS’s untyped storage.

RawFns

Helper trait that is auto-implemented for all Clone-able types. Provides easy access to drop and clone funcitons for raw pointers.

TypedEcsData

Helper trait that is auto-implemented for anything that may be stored in the ECS’s typed storage.